feat: 비밀번호를 입력하여 게임 방을 삭제할 수 있도록 수정

+++ src/main/java/chess/controller/ChessApiController.java
@@ -39,8 +39,8 @@ public class ChessApiController {
     }
 
     @DeleteMapping("/game/{id}")
-    public void deleteGame(@PathVariable Long id) {
-        chessService.deleteByGameId(id);
+    public void deleteGame(@PathVariable Long id, @RequestBody String confirmPwd) {
+        chessService.deleteByGameId(confirmPwd, id);
     }
 
     @PostMapping("/game/{id}/move")
